moved icons folder
