+++ pages/events/search.js
@@ -2,7 +2,7 @@ import { useRouter } from "next/router";
 import EventList from "@/components/EventList";
 import Button from "@/components/Button";
 import Heading from "@/components/ui/Heading";
-import ChevronLeft from "@/components/ui/icons/ChevronLeft";
+import ChevronLeft from "@/components/icons/ChevronLeft";
 export default function Search() {
   const router = useRouter();
   const { q } = router.query;
